requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
pandas>=2.2.0
numpy>=1.26.0
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
cloud-sql-python-connector[pg8000]>=1.4.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
sentence-transformers>=2.2.0
plotly>=5.17.0
openpyxl>=3.1.0
jieba>=0.42.1
nltk>=3.8.1
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0